requirements.txt
# requirements.txt
streamlit>=1.20.0
pandas>=1.3.5
numpy>=1.21.6
opencv-python>=4.5.5.64
pytesseract>=0.3.10
pdf2image>=1.16.0
docx2txt>=0.8
PyMuPDF>=1.20.0
spacy>=3.4.0
spacy-transformers>=1.1.8
langdetect>=1.0.9
pyyaml>=6.0
xlsxwriter>=3.0.3
rapidfuzz>=2.13.7
psutil>=5.9.0
tqdm>=4.64.1